Mange multi image by admin

+++ core/shop/models.py
@@ -2,18 +2,19 @@ from django.db import models
 from decimal import Decimal
 from django.core.validators import MinValueValidator, MaxValueValidator
 from django.urls import reverse
-from django.db.models.signals import pre_delete
-from django.dispatch import receiver
-from django.utils.translation import get_language
 from parler.models import TranslatableModel, TranslatedFields
 from core.utils.liara_upload import upload_to_liara, delete_from_liara
+from django.conf import settings
+from django.utils.translation import get_language
+from django.db.models.signals import pre_delete
+from django.dispatch import receiver
 
 
 class ProductStatusType(models.IntegerChoices):
-    publish = 1, ("نمایش")
-    draft = 2, ("عدم نمایش")
-
-
+    publish = 1, ("Show")
+    draft = 2, ("Hide")
+    
+    
 class ProductCategoryModel(TranslatableModel):
     translations = TranslatedFields(
         title=models.CharField(max_length=255),
@@ -28,8 +29,8 @@ class ProductCategoryModel(TranslatableModel):
 
 class ProductModel(TranslatableModel):
     user = models.ForeignKey("accounts.User", on_delete=models.PROTECT)
-    category = models.ManyToManyField(ProductCategoryModel)
-
+    category = models.ManyToManyField("ProductCategoryModel")
+    
     translations = TranslatedFields(
         title=models.CharField(max_length=255),
         slug=models.SlugField(allow_unicode=True),
@@ -44,53 +45,27 @@ class ProductModel(TranslatableModel):
     price = models.DecimalField(default=0, max_digits=10, decimal_places=0)
     discount_percent = models.IntegerField(default=0, validators=[MinValueValidator(0), MaxValueValidator(100)])
     avg_rate = models.FloatField(default=0.0)
-
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
 
-    class Meta:
-        ordering = ["-created_at"]
-
-    def __str__(self):
-        return self.safe_translation_getter("title", any_language=True)
-
     def save(self, *args, **kwargs):
-        super().save(*args, **kwargs)  # Save first to get an ID
-        # Upload main image to Liara if not uploaded yet
-        if self.image and not self.image_url:
-            uploaded_url = upload_to_liara(self.image, f"product-main-{self.id}.jpg", folder="product/img")
-            if uploaded_url:
-                self.image_url = uploaded_url
-                super().save(update_fields=['image_url'])
-
-    def get_price(self):
-        discount_amount = self.price * Decimal(self.discount_percent / 100)
-        final_price = self.price - discount_amount
-        return round(final_price)
-
-    def is_discounted(self):
-        return self.discount_percent != 0
-
-    def is_published(self):
-        return self.status == ProductStatusType.publish.value
+        super().save(*args, **kwargs)
+        if self.image:
+            if settings.LIARA_UPLOAD_ENABLED:
+                uploaded_url = upload_to_liara(self.image, f"product-main-{self.id}.jpg", folder="product/img")
+                if uploaded_url:
+                    self.image_url = uploaded_url
+                    super().save(update_fields=['image_url'])
+            else:  # Local
+                if not self.image_url:
+                    self.image_url = self.image.url
+                    super().save(update_fields=['image_url'])
 
     def get_absolute_url(self):
-        lang = get_language()
-        slug = self.safe_translation_getter("slug", language_code=lang, any_language=True)
+        slug = self.safe_translation_getter("slug", any_language=True)
         if not slug:
-            return reverse("shop:product-detail", kwargs={"slug": self.id})
+            return reverse("shop:product-detail", kwargs={"slug": str(self.id)})
         return reverse("shop:product-detail", kwargs={"slug": slug})
-
-    def delete_main_image(self):
-        if self.image_url:
-            key = "/".join(self.image_url.split('/')[-2:])
-            delete_from_liara(key)
-            self.image_url = None
-        if self.image:
-            self.image.delete(save=False)
-            self.image = None
-        self.save()
-
 
 class ProductImageModel(models.Model):
     product = models.ForeignKey(ProductModel, on_delete=models.CASCADE, related_name="extra_images")
@@ -101,20 +76,17 @@ class ProductImageModel(models.Model):
     updated_at = models.DateTimeField(auto_now=True)
 
     def save(self, *args, **kwargs):
-        super().save(*args, **kwargs)  # Save first to get an ID
-        # Upload extra image to Liara if not uploaded yet
-        if self.file and not self.url:
-            uploaded_url = upload_to_liara(self.file, f"product-extra-{self.id}.jpg", folder="product/extra-img")
-            if uploaded_url:
-                self.url = uploaded_url
-                super().save(update_fields=['url'])
-
-    def delete_image(self):
-        if self.url:
-            key = "/".join(self.url.split('/')[-2:])
-            delete_from_liara(key)
+        super().save(*args, **kwargs)
         if self.file:
-            self.file.delete(save=False)
+            if settings.LIARA_UPLOAD_ENABLED:
+                uploaded_url = upload_to_liara(self.file, f"product-extra-{self.id}.jpg", folder="product/extra-img")
+                if uploaded_url:
+                    self.url = uploaded_url
+                    super().save(update_fields=['url'])
+            else:  # Local
+                if not self.url:
+                    self.url = self.file.url
+                    super().save(update_fields=['url'])
 
 
 class WishlistProductModel(models.Model):
@@ -127,7 +99,10 @@ class WishlistProductModel(models.Model):
 
 @receiver(pre_delete, sender=ProductModel)
 def delete_product_images(sender, instance, **kwargs):
-    instance.delete_main_image()
+    if hasattr(instance, 'delete_main_image'):
+        instance.delete_main_image()
     for img in instance.extra_images.all():
-        img.delete_image()
+        if hasattr(img, 'delete_image'):
+            img.delete_image()
         img.delete()
+
